refactoring: now we will use these classes to test all table add columns migrations, not only `veTable` migration #8651 only:uaefi
