+++ java_console/ui/src/test/java/com/rusefi/maintenance/migration/default_migration/DefaultTuneMigratorTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import javax.xml.bind.JAXBException;
 
 import static com.rusefi.maintenance.migration.default_migration.DefaultTestTuneMigrationContext.*;
-import static com.rusefi.maintenance.migration.ve_table_extension.VeTableExtensionTestTuneMigrationContext.VE_RPM_BINS_FIELD_NAME;
-import static com.rusefi.maintenance.migration.ve_table_extension.VeTableExtensionTestTuneMigrationContext.VE_TABLE_FIELD_NAME;
+import static com.rusefi.maintenance.migration.table_add_columns_migration.TableAddColumnsMigrationContext.VE_RPM_BINS_FIELD_NAME;
+import static com.rusefi.maintenance.migration.table_add_columns_migration.TableAddColumnsMigrationContext.VE_TABLE_FIELD_NAME;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class DefaultTuneMigratorTest {
