+++ java_console/ui/src/test/java/com/rusefi/maintenance/migration/table_add_columns_migration/test_data/prev_calibrations.msq
